Move server-related utils to server
- namely: get_artifacts_from_payload, validate_request_size
- it avoids circular imports
- the location makes much more sense, since these functions use
fastapi, models, exceptions and other utilities from utils module
